refactor(rag): extract VectorStore provider from RagService
Move the full LanceDB table-management layer (connection handling,
table cache/locks, indexDocument/search/listDocuments/deleteBySource/
optimizeTable/deleteAllBySourceType/validateTableProvider, and their
OnModuleInit/OnModuleDestroy lifecycle) out of RagService into a new
VectorStore provider. RagService now delegates to VectorStore for all
of these and keeps only ragRepository/txManager/graphStore/
incrementalDiff plus importGraphify/importGraphifyFull orchestration.

Pure behavior-preserving extraction: no forwardRef removed, no
RagModule imports/exports change, RagService's public API unchanged.
Sets up the IncrementalGraphDiffService<->RagService forwardRef cycle
to be broken in a follow-up task by having IncrementalGraphDiffService
depend on VectorStore instead of RagService.

Split rag.service.spec.ts accordingly: LanceDB-touching test cases
moved to the new vector-store.service.spec.ts, RagService's spec keeps
only the free-function unit tests and the importGraphify orchestration
suite.
