+++ apps/api/src/rag/rag.service.spec.ts
@@ -1,28 +1,3 @@
-// Prevent LanceDB native module from loading — it holds open handles in tests
-jest.mock('@lancedb/lancedb', () => ({
-  connect: jest.fn().mockResolvedValue({
-    tableNames: jest.fn().mockResolvedValue([]),
-    createTable: jest.fn().mockResolvedValue({
-      delete: jest.fn().mockResolvedValue(undefined),
-      createIndex: jest.fn().mockResolvedValue(undefined),
-      add: jest.fn().mockResolvedValue(undefined),
-      query: jest.fn().mockReturnValue({ limit: jest.fn().mockReturnThis(), toArray: jest.fn().mockResolvedValue([]) }),
-      countRows: jest.fn().mockResolvedValue(0),
-    }),
-    openTable: jest.fn().mockResolvedValue({
-      delete: jest.fn().mockResolvedValue(undefined),
-      createIndex: jest.fn().mockResolvedValue(undefined),
-      add: jest.fn().mockResolvedValue(undefined),
-      query: jest.fn().mockReturnValue({ limit: jest.fn().mockReturnThis(), toArray: jest.fn().mockResolvedValue([]) }),
-      countRows: jest.fn().mockResolvedValue(0),
-      search: jest.fn().mockResolvedValue([]),
-      vectorSearch: jest.fn().mockReturnValue({ distanceType: jest.fn().mockReturnThis(), limit: jest.fn().mockReturnThis(), toArray: jest.fn().mockResolvedValue([]) }),
-      optimize: jest.fn().mockResolvedValue(undefined),
-    }),
-  }),
-  Index: { fts: jest.fn().mockReturnValue({}) },
-}));
-
 import { IRagConfig } from '../config/rag.config';
 import {
   RagService,
@@ -31,9 +6,7 @@ import {
   getSimilarityTier,
   getVerdict,
 } from './rag.service';
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-type AnyRecord = Record<string, any>;
+import { VectorStore } from './vector-store.service';
 
 function makeRagConfig(overrides: Partial<IRagConfig> = {}): IRagConfig {
   return {
@@ -54,29 +27,6 @@ function makeRagConfig(overrides: Partial<IRagConfig> = {}): IRagConfig {
     ...overrides,
   };
 }
-
-function makeMockRecord(id: string, content: string): AnyRecord {
-  return {
-    id,
-    source: 'ticket',
-    source_id: id,
-    content,
-    vector: [],
-    metadata: '{}',
-    created_at: new Date().toISOString(),
-    provider: 'ollama',
-    model: 'nomic-embed-text',
-  };
-}
-
-const mockRagConfigForSearch = makeRagConfig();
-
-const mockEmbeddingServiceForSearch = {
-  embed: jest.fn().mockResolvedValue(new Float32Array(384).fill(0)),
-  getDimensions: jest.fn().mockReturnValue(384),
-  provider: 'ollama',
-  model: 'nomic-embed-text',
-};
 
 describe('simpleFtsScore', () => {
   it('returns 0 for empty query', () => {
@@ -206,514 +156,10 @@ describe('getVerdict', () => {
   });
 });
 
-describe('RagService lifecycle', () => {
-  it('closes LanceDB connection on module destroy', () => {
-    const ragService = new RagService(makeRagConfig());
-    const closeSpy = jest.fn();
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).db = { close: closeSpy };
-
-    ragService.onModuleDestroy();
-
-    expect(closeSpy).toHaveBeenCalledTimes(1);
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    expect((ragService as any).db).toBeNull();
-  });
-});
-
-describe('RagService.getOrCreateTable — FTS index creation', () => {
-  const mockRagConfig = makeRagConfig();
-
-  it('calls table.createIndex with FTS config when lanceAvailable is true', async () => {
-    const ragService = new RagService(mockRagConfig);
-    const createIndexSpy = jest.fn().mockResolvedValue(undefined);
-    const deleteSpy = jest.fn().mockResolvedValue(undefined);
-
-    const mockDb = {
-      tableNames: jest.fn().mockResolvedValue([]),
-      createTable: jest.fn().mockResolvedValue({
-        delete: deleteSpy,
-        createIndex: createIndexSpy,
-      }),
-    };
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).db = mockDb;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-
-    await ragService.getOrCreateTable('test-project');
-
-    expect(createIndexSpy).toHaveBeenCalledWith(
-      'content',
-      expect.objectContaining({
-        replace: false,
-        config: expect.anything(),
-      }),
-    );
-  });
-
-  it('does not call table.createIndex when lanceAvailable is false', async () => {
-    const ragService = new RagService(mockRagConfig);
-    ragService.onModuleInit();
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = false;
-
-    await ragService.getOrCreateTable('test-project');
-
-    // Verify no error thrown and in-memory table returned
-    const table = await ragService.getOrCreateTable('test-project');
-    expect(table).toBeDefined();
-  });
-
-  it('logs warning and does not throw when createIndex rejects', async () => {
-    const ragService = new RagService(mockRagConfig);
-    const loggerSpy = jest.spyOn(ragService['logger'], 'warn');
-    const createIndexError = new Error('Index already exists');
-    const createIndexSpy = jest.fn().mockRejectedValue(createIndexError);
-    const deleteSpy = jest.fn().mockResolvedValue(undefined);
-
-    const mockDb = {
-      tableNames: jest.fn().mockResolvedValue([]),
-      createTable: jest.fn().mockResolvedValue({
-        delete: deleteSpy,
-        createIndex: createIndexSpy,
-      }),
-    };
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).db = mockDb;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-
-    // Should not throw
-    await expect(ragService.getOrCreateTable('test-project')).resolves.toBeDefined();
-    expect(loggerSpy).toHaveBeenCalled();
-    expect(loggerSpy.mock.calls[0][0]).toContain('FTS index');
-  });
-
-  it('serializes concurrent calls for the same project so the table is only created once', async () => {
-    const ragService = new RagService(mockRagConfig);
-
-    let resolveCreateTable = (_table: unknown): void => {};
-    const createTablePromise = new Promise((resolve) => {
-      resolveCreateTable = resolve;
-    });
-    const createTableSpy = jest.fn().mockReturnValue(createTablePromise);
-
-    const mockDb = {
-      tableNames: jest.fn().mockResolvedValue([]),
-      createTable: createTableSpy,
-    };
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).db = mockDb;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-
-    const first = ragService.getOrCreateTable('concurrent-project');
-    const second = ragService.getOrCreateTable('concurrent-project');
-
-    // Let both calls' preceding awaits (validateProjectId, connect, tableNames) flush
-    // before asserting createTable was only invoked once.
-    await new Promise((resolve) => setImmediate(resolve));
-    await new Promise((resolve) => setImmediate(resolve));
-
-    expect(createTableSpy).toHaveBeenCalledTimes(1);
-
-    resolveCreateTable({
-      delete: jest.fn().mockResolvedValue(undefined),
-      createIndex: jest.fn().mockResolvedValue(undefined),
-    });
-
-    const [firstTable, secondTable] = await Promise.all([first, second]);
-    expect(firstTable).toBe(secondTable);
-    expect(createTableSpy).toHaveBeenCalledTimes(1);
-  });
-});
-
-describe('RagService — write mutex serialization (LanceDB has no built-in concurrent-writer lock)', () => {
-  const mockRagConfig = makeRagConfig();
-
-  const mockEmbeddingService = {
-    embed: jest.fn().mockResolvedValue([0.1, 0.2, 0.3]),
-    providerName: 'test-provider',
-    modelName: 'test-model',
-    dimensions: 3,
-  };
-
-  it('serializes concurrent indexDocument writes against the same project table', async () => {
-    const ragService = new RagService(mockRagConfig, mockEmbeddingService as never);
-
-    let resolveFirstAdd: () => void = () => {};
-    const firstAddPromise = new Promise<void>((resolve) => {
-      resolveFirstAdd = resolve;
-    });
-    const addSpy = jest.fn().mockReturnValueOnce(firstAddPromise).mockResolvedValueOnce(undefined);
-    const mockTable = { add: addSpy };
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    const first = ragService.indexDocument('test-project', {
-      source: 'ticket',
-      sourceId: 'ticket-001',
-      content: 'first',
-      metadata: {},
-    });
-    const second = ragService.indexDocument('test-project', {
-      source: 'ticket',
-      sourceId: 'ticket-002',
-      content: 'second',
-      metadata: {},
-    });
-
-    // Let both calls' preceding awaits (validateProjectId, embed) flush before
-    // asserting the second write hasn't started — it must wait on the first.
-    await new Promise((resolve) => setImmediate(resolve));
-    await new Promise((resolve) => setImmediate(resolve));
-
-    expect(addSpy).toHaveBeenCalledTimes(1);
-
-    resolveFirstAdd();
-    await Promise.all([first, second]);
-
-    expect(addSpy).toHaveBeenCalledTimes(2);
-  });
-
-  it('never runs indexDocument and deleteBySource writes concurrently on the same table', async () => {
-    // indexDocument and deleteBySource have different numbers of awaits before
-    // reaching the write lock, so which one acquires it first isn't guaranteed —
-    // what the mutex guarantees is that only one write is ever in flight at a time.
-    const ragService = new RagService(mockRagConfig, mockEmbeddingService as never);
-
-    let resolveAdd: () => void = () => {};
-    const addPromise = new Promise<void>((resolve) => {
-      resolveAdd = resolve;
-    });
-    const addSpy = jest.fn().mockReturnValue(addPromise);
-
-    let resolveDelete: () => void = () => {};
-    const deletePromiseGate = new Promise<void>((resolve) => {
-      resolveDelete = resolve;
-    });
-    const deleteSpy = jest.fn().mockReturnValue(deletePromiseGate);
-
-    const mockTable = { add: addSpy, delete: deleteSpy };
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    const indexPromise = ragService.indexDocument('test-project', {
-      source: 'ticket',
-      sourceId: 'ticket-001',
-      content: 'first',
-      metadata: {},
-    });
-    const deletePromise = ragService.deleteBySource('test-project', 'ticket-001');
-
-    for (let i = 0; i < 5; i++) {
-      await new Promise((resolve) => setImmediate(resolve));
-    }
-
-    // Only one of the two writes may be in flight at a time.
-    expect(addSpy.mock.calls.length + deleteSpy.mock.calls.length).toBe(1);
-
-    resolveAdd();
-    resolveDelete();
-    await Promise.all([indexPromise, deletePromise]);
-
-    expect(addSpy).toHaveBeenCalledTimes(1);
-    expect(deleteSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it('does not let one project table lock block writes to a different project table', async () => {
-    const ragService = new RagService(mockRagConfig, mockEmbeddingService as never);
-
-    let resolveAddA: () => void = () => {};
-    const addAPromise = new Promise<void>((resolve) => {
-      resolveAddA = resolve;
-    });
-    const addASpy = jest.fn().mockReturnValue(addAPromise);
-    const addBSpy = jest.fn().mockResolvedValue(undefined);
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([
-      ['project_project-a', { add: addASpy }],
-      ['project_project-b', { add: addBSpy }],
-    ]);
-
-    const pendingA = ragService.indexDocument('project-a', {
-      source: 'ticket',
-      sourceId: 'a-1',
-      content: 'a',
-      metadata: {},
-    });
-    await ragService.indexDocument('project-b', {
-      source: 'ticket',
-      sourceId: 'b-1',
-      content: 'b',
-      metadata: {},
-    });
-
-    expect(addBSpy).toHaveBeenCalledTimes(1);
-
-    resolveAddA();
-    await pendingA;
-  });
-});
-
 describe('reciprocalRankFusion — export (US-003-2 AC-3)', () => {
   it('is exported from rag.service.ts', () => {
     expect(reciprocalRankFusion).toBeDefined();
     expect(typeof reciprocalRankFusion).toBe('function');
-  });
-});
-
-describe('RagService.search — native FTS path (US-003-2)', () => {
-  function makeTableWithFts(
-    allRows: AnyRecord[],
-    ftsRows: AnyRecord[],
-    vectorRows: AnyRecord[] = [],
-  ) {
-    return {
-      countRows: jest.fn().mockResolvedValue(allRows.length || 1),
-      query: jest.fn().mockReturnValue({
-        limit: jest.fn().mockReturnThis(),
-        toArray: jest.fn().mockResolvedValue(allRows),
-      }),
-      search: jest.fn().mockResolvedValue(ftsRows),
-      vectorSearch: jest.fn().mockReturnValue({
-        distanceType: jest.fn().mockReturnThis(),
-        limit: jest.fn().mockReturnThis(),
-        toArray: jest.fn().mockResolvedValue(vectorRows),
-      }),
-    };
-  }
-
-  it('calls table.search(query, "fts", "content") when lanceAvailable is true', async () => {
-    const ragService = new RagService(
-      mockRagConfigForSearch,
-      mockEmbeddingServiceForSearch as never,
-    );
-
-    const ftsRow = makeMockRecord('fts-doc-1', 'authentication error in service');
-    const mockTable = makeTableWithFts([ftsRow], [ftsRow]);
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    await ragService.search('test-project', 'authentication error');
-
-    expect(mockTable.search).toHaveBeenCalledWith('authentication error', 'fts', 'content');
-  });
-
-  it('does not call table.search() when lanceAvailable is false', async () => {
-    const ragService = new RagService(
-      mockRagConfigForSearch,
-      mockEmbeddingServiceForSearch as never,
-    );
-
-    const doc = makeMockRecord('doc-1', 'auth keyword content here');
-    const mockTable = makeTableWithFts([doc], [doc]);
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = false;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    await ragService.search('test-project', 'auth content');
-
-    expect(mockTable.search).not.toHaveBeenCalled();
-  });
-
-  it('scores native FTS results by reciprocal position 1/(i+1)', async () => {
-    const ragService = new RagService(
-      mockRagConfigForSearch,
-      mockEmbeddingServiceForSearch as never,
-    );
-
-    const ftsRows = [
-      makeMockRecord('fts-1', 'first result document'),
-      makeMockRecord('fts-2', 'second result document'),
-      makeMockRecord('fts-3', 'third result document'),
-    ];
-
-    // No vector results — FTS scores drive final output
-    const mockTable = makeTableWithFts(ftsRows, ftsRows, []);
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    const result = await ragService.search('test-project', 'result', 3);
-
-    expect(result.results).toHaveLength(3);
-    // FTS result at index 0 → score = 1/(0+1) = 1.0
-    expect(result.results[0].score).toBeCloseTo(1 / (0 + 1));
-    // FTS result at index 1 → score = 1/(1+1) = 0.5
-    expect(result.results[1].score).toBeCloseTo(1 / (1 + 1));
-    // FTS result at index 2 → score = 1/(2+1) ≈ 0.333
-    expect(result.results[2].score).toBeCloseTo(1 / (2 + 1));
-  });
-
-  it('adds ftsRows to recordMap so records unique to native FTS appear in results', async () => {
-    const ragService = new RagService(
-      mockRagConfigForSearch,
-      mockEmbeddingServiceForSearch as never,
-    );
-
-    // fts-only-doc is returned by native FTS but not in the allRows table scan
-    const scannedDoc = makeMockRecord('scanned-doc', 'some generic content');
-    const ftsOnlyDoc = makeMockRecord('fts-only-doc', 'authentication failure critical');
-
-    const mockTable = makeTableWithFts(
-      [scannedDoc],          // allRows: only scanned-doc
-      [ftsOnlyDoc, scannedDoc], // ftsRows: fts-only-doc is the top FTS hit
-      [],                    // vectorRows: empty
-    );
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    const result = await ragService.search('test-project', 'authentication failure', 5);
-
-    const ids = result.results.map((r) => r.id);
-    // fts-only-doc must appear in results because ftsRows were added to recordMap
-    expect(ids).toContain('fts-only-doc');
-  });
-});
-
-describe('RagService.search — in-memory FTS fallback path (US-003-3)', () => {
-  function makeTableWithRejectedSearch(allRows: AnyRecord[]) {
-    return {
-      countRows: jest.fn().mockResolvedValue(allRows.length || 1),
-      query: jest.fn().mockReturnValue({
-        limit: jest.fn().mockReturnThis(),
-        toArray: jest.fn().mockResolvedValue(allRows),
-      }),
-      search: jest.fn().mockRejectedValue(new Error('tantivy index not ready')),
-      vectorSearch: jest.fn().mockReturnValue({
-        distanceType: jest.fn().mockReturnThis(),
-        limit: jest.fn().mockReturnThis(),
-        toArray: jest.fn().mockResolvedValue([]),
-      }),
-    };
-  }
-
-  it('when lanceAvailable=true and table.search() rejects, falls back to simpleFtsScore and returns non-empty results', async () => {
-    const ragService = new RagService(
-      mockRagConfigForSearch,
-      mockEmbeddingServiceForSearch as never,
-    );
-
-    const matchingDoc = makeMockRecord('matching-doc', 'authentication error occurred in service');
-    const mockTable = makeTableWithRejectedSearch([matchingDoc]);
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    const result = await ragService.search('test-project', 'authentication error', 5);
-
-    // table.search was called (and rejected)
-    expect(mockTable.search).toHaveBeenCalled();
-    // but results must still be populated via simpleFtsScore fallback
-    expect(result.results.length).toBeGreaterThan(0);
-    expect(result.results[0].id).toBe('matching-doc');
-  });
-
-  it('when lanceAvailable=true and table.search() rejects, logs a warning', async () => {
-    const ragService = new RagService(
-      mockRagConfigForSearch,
-      mockEmbeddingServiceForSearch as never,
-    );
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const loggerWarnSpy = jest.spyOn((ragService as any).logger, 'warn');
-
-    const doc = makeMockRecord('doc-1', 'some matching content');
-    const mockTable = makeTableWithRejectedSearch([doc]);
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    await ragService.search('test-project', 'matching content', 5);
-
-    const warnCalls = loggerWarnSpy.mock.calls.map((c) => String(c[0]));
-    expect(warnCalls.some((msg) => msg.toLowerCase().includes('fts'))).toBe(true);
-  });
-
-  it('when lanceAvailable=true and table.search() rejects, fallback scores match simpleFtsScore output', async () => {
-    const ragService = new RagService(
-      mockRagConfigForSearch,
-      mockEmbeddingServiceForSearch as never,
-    );
-
-    const fullMatchDoc = makeMockRecord('full-match', 'authentication error service crash');
-    const partialMatchDoc = makeMockRecord('partial-match', 'authentication only partial');
-    const noMatchDoc = makeMockRecord('no-match', 'database schema migration rollback');
-
-    const mockTable = makeTableWithRejectedSearch([fullMatchDoc, partialMatchDoc, noMatchDoc]);
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    const result = await ragService.search('test-project', 'authentication error service', 5);
-
-    // no-match doc should not appear — simpleFtsScore returns 0 for it
-    const ids = result.results.map((r) => r.id);
-    expect(ids).not.toContain('no-match');
-    // full-match should score higher than partial-match
-    const fullIdx = ids.indexOf('full-match');
-    const partialIdx = ids.indexOf('partial-match');
-    expect(fullIdx).toBeGreaterThanOrEqual(0);
-    expect(partialIdx).toBeGreaterThanOrEqual(0);
-    expect(fullIdx).toBeLessThan(partialIdx);
-  });
-
-  it('when lanceAvailable=false, uses simpleFtsScore without calling table.search()', async () => {
-    const ragService = new RagService(
-      mockRagConfigForSearch,
-      mockEmbeddingServiceForSearch as never,
-    );
-
-    const doc = makeMockRecord('doc-in-memory', 'authentication error keyword content');
-    const mockTable = {
-      countRows: jest.fn().mockResolvedValue(1),
-      query: jest.fn().mockReturnValue({
-        limit: jest.fn().mockReturnThis(),
-        toArray: jest.fn().mockResolvedValue([doc]),
-      }),
-      search: jest.fn().mockResolvedValue([]),
-      vectorSearch: jest.fn().mockReturnValue({
-        distanceType: jest.fn().mockReturnThis(),
-        limit: jest.fn().mockReturnThis(),
-        toArray: jest.fn().mockResolvedValue([]),
-      }),
-    };
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = false;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    const result = await ragService.search('test-project', 'authentication error', 5);
-
-    expect(mockTable.search).not.toHaveBeenCalled();
-    expect(result.results.length).toBeGreaterThan(0);
-    expect(result.results[0].id).toBe('doc-in-memory');
   });
 });
 
@@ -728,314 +174,6 @@ describe('simpleFtsScore — export (US-003-3 AC-3)', () => {
   });
 });
 
-describe('RagService — onFirstAccess and onDestroy lifecycle hooks (US-003-5)', () => {
-  const mockRagConfig = makeRagConfig();
-
-  function makeMockDb(tableExists: boolean) {
-    const mockTable = {
-      delete: jest.fn().mockResolvedValue(undefined),
-      createIndex: jest.fn().mockResolvedValue(undefined),
-    };
-    return {
-      tableNames: jest.fn().mockResolvedValue(tableExists ? ['project_test-project'] : []),
-      openTable: jest.fn().mockResolvedValue(mockTable),
-      createTable: jest.fn().mockResolvedValue(mockTable),
-      mockTable,
-    };
-  }
-
-  describe('getOrCreateTable — onFirstAccess', () => {
-    it('calls optimizeStrategy.onFirstAccess(projectId, table) when lanceAvailable is true', async () => {
-      const ragService = new RagService(mockRagConfig);
-      const onFirstAccessSpy = jest.fn();
-      const mockStrategy = { onFirstAccess: onFirstAccessSpy, onInsert: jest.fn(), onDestroy: jest.fn() };
-      const { mockTable, ...mockDb } = makeMockDb(false);
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (ragService as any).db = mockDb;
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (ragService as any).lanceAvailable = true;
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (ragService as any).optimizeStrategy = mockStrategy;
-
-      await ragService.getOrCreateTable('test-project');
-
-      expect(onFirstAccessSpy).toHaveBeenCalledWith('test-project', mockTable);
-    });
-
-    it('calls optimizeStrategy.onFirstAccess exactly once per projectId even when called multiple times', async () => {
-      const ragService = new RagService(mockRagConfig);
-      const onFirstAccessSpy = jest.fn();
-      const mockStrategy = { onFirstAccess: onFirstAccessSpy, onInsert: jest.fn(), onDestroy: jest.fn() };
-      const { mockTable, ...mockDb } = makeMockDb(false);
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (ragService as any).db = mockDb;
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (ragService as any).lanceAvailable = true;
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (ragService as any).optimizeStrategy = mockStrategy;
-
-      await ragService.getOrCreateTable('test-project');
-      await ragService.getOrCreateTable('test-project');
-      await ragService.getOrCreateTable('test-project');
-
-      expect(onFirstAccessSpy).toHaveBeenCalledTimes(1);
-    });
-
-    it('calls optimizeStrategy.onFirstAccess once per distinct projectId', async () => {
-      const ragService = new RagService(mockRagConfig);
-      const onFirstAccessSpy = jest.fn();
-      const mockStrategy = { onFirstAccess: onFirstAccessSpy, onInsert: jest.fn(), onDestroy: jest.fn() };
-
-      const mockTableA = { delete: jest.fn().mockResolvedValue(undefined), createIndex: jest.fn().mockResolvedValue(undefined) };
-      const mockTableB = { delete: jest.fn().mockResolvedValue(undefined), createIndex: jest.fn().mockResolvedValue(undefined) };
-      const mockDb = {
-        tableNames: jest.fn().mockResolvedValue([]),
-        createTable: jest.fn()
-          .mockResolvedValueOnce(mockTableA)
-          .mockResolvedValueOnce(mockTableB),
-      };
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (ragService as any).db = mockDb;
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (ragService as any).lanceAvailable = true;
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (ragService as any).optimizeStrategy = mockStrategy;
-
-      await ragService.getOrCreateTable('project-a');
-      await ragService.getOrCreateTable('project-b');
-      // Second call for project-a — must NOT trigger another onFirstAccess
-      await ragService.getOrCreateTable('project-a');
-
-      expect(onFirstAccessSpy).toHaveBeenCalledTimes(2);
-      expect(onFirstAccessSpy).toHaveBeenCalledWith('project-a', mockTableA);
-      expect(onFirstAccessSpy).toHaveBeenCalledWith('project-b', mockTableB);
-    });
-
-    it('does not call optimizeStrategy.onFirstAccess when lanceAvailable is false', async () => {
-      const ragService = new RagService(mockRagConfig);
-      const onFirstAccessSpy = jest.fn();
-      const mockStrategy = { onFirstAccess: onFirstAccessSpy, onInsert: jest.fn(), onDestroy: jest.fn() };
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (ragService as any).lanceAvailable = false;
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (ragService as any).optimizeStrategy = mockStrategy;
-
-      await ragService.getOrCreateTable('test-project');
-
-      expect(onFirstAccessSpy).not.toHaveBeenCalled();
-    });
-
-    it('does not call optimizeStrategy.onFirstAccess when optimizeStrategy is not injected', async () => {
-      const ragService = new RagService(mockRagConfig);
-      const { mockTable, ...mockDb } = makeMockDb(false);
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (ragService as any).db = mockDb;
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (ragService as any).lanceAvailable = true;
-      // optimizeStrategy is undefined (not injected)
-
-      await expect(ragService.getOrCreateTable('test-project')).resolves.toBeDefined();
-    });
-  });
-
-  describe('onModuleDestroy — onDestroy', () => {
-    it('calls optimizeStrategy.onDestroy() during onModuleDestroy', async () => {
-      const ragService = new RagService(mockRagConfig);
-      const onDestroySpy = jest.fn().mockResolvedValue(undefined);
-      const mockStrategy = { onFirstAccess: jest.fn(), onInsert: jest.fn(), onDestroy: onDestroySpy };
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (ragService as any).optimizeStrategy = mockStrategy;
-
-      await ragService.onModuleDestroy();
-
-      expect(onDestroySpy).toHaveBeenCalledTimes(1);
-    });
-
-    it('calls optimizeStrategy.onDestroy() even when no LanceDB connection exists', async () => {
-      const ragService = new RagService(mockRagConfig);
-      const onDestroySpy = jest.fn().mockResolvedValue(undefined);
-      const mockStrategy = { onFirstAccess: jest.fn(), onInsert: jest.fn(), onDestroy: onDestroySpy };
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (ragService as any).db = null;
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (ragService as any).optimizeStrategy = mockStrategy;
-
-      await ragService.onModuleDestroy();
-
-      expect(onDestroySpy).toHaveBeenCalledTimes(1);
-    });
-
-    it('does not throw when optimizeStrategy is not injected', async () => {
-      const ragService = new RagService(mockRagConfig);
-      // optimizeStrategy is undefined (not injected)
-
-      await expect(ragService.onModuleDestroy()).resolves.toBeUndefined();
-    });
-  });
-});
-
-describe('RagService.indexDocument — onInsert Strategy Hook (US-003-4)', () => {
-  const mockRagConfig = makeRagConfig();
-
-  const mockEmbeddingService = {
-    embed: jest.fn().mockResolvedValue([0.1, 0.2, 0.3]),
-    providerName: 'test-provider',
-    modelName: 'test-model',
-    dimensions: 3,
-  };
-
-  it('calls optimizeStrategy.onInsert(projectId, table) after table.add() when lanceAvailable is true', async () => {
-    const ragService = new RagService(mockRagConfig, mockEmbeddingService as never);
-    const onInsertSpy = jest.fn().mockResolvedValue(undefined);
-    const mockStrategy = { onInsert: onInsertSpy } as unknown as never;
-    const mockTable = { add: jest.fn().mockResolvedValue(undefined) };
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).optimizeStrategy = mockStrategy;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    await ragService.indexDocument('test-project', {
-      source: 'ticket',
-      sourceId: 'ticket-001',
-      content: 'Test document content',
-      metadata: { ref: 'TEST-1' },
-    });
-
-    expect(onInsertSpy).toHaveBeenCalledWith('test-project', mockTable);
-  });
-
-  it('does not call optimizeStrategy.onInsert() when lanceAvailable is false', async () => {
-    const ragService = new RagService(mockRagConfig, mockEmbeddingService as never);
-    const onInsertSpy = jest.fn().mockResolvedValue(undefined);
-    const mockStrategy = { onInsert: onInsertSpy } as unknown as never;
-    const mockTable = { add: jest.fn().mockResolvedValue(undefined) };
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = false;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).optimizeStrategy = mockStrategy;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    await ragService.indexDocument('test-project', {
-      source: 'ticket',
-      sourceId: 'ticket-001',
-      content: 'Test document content',
-      metadata: { ref: 'TEST-1' },
-    });
-
-    expect(onInsertSpy).not.toHaveBeenCalled();
-  });
-});
-
-describe('RagService.optimizeTable (US-004)', () => {
-  it('calls table.optimize() when lanceAvailable is true', async () => {
-    const ragService = new RagService(makeRagConfig());
-    const mockTable = { optimize: jest.fn().mockResolvedValue(undefined) };
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    jest.spyOn(ragService as any, 'getOrCreateTable').mockResolvedValue(mockTable);
-
-    await ragService.optimizeTable('test-project');
-
-    expect(mockTable.optimize).toHaveBeenCalledTimes(1);
-  });
-
-  it('does not call table.optimize() when lanceAvailable is false', async () => {
-    const ragService = new RagService(makeRagConfig());
-    const mockTable = { optimize: jest.fn().mockResolvedValue(undefined) };
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = false;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    jest.spyOn(ragService as any, 'getOrCreateTable').mockResolvedValue(mockTable);
-
-    await ragService.optimizeTable('test-project');
-
-    expect(mockTable.optimize).not.toHaveBeenCalled();
-  });
-});
-
-describe('RagService.deleteAllBySourceType (US-002)', () => {
-  it('AC1: deletes all records where source = sourceType and returns count', async () => {
-    const ragService = new RagService(makeRagConfig());
-    const mockTable = {
-      countRows: jest.fn().mockResolvedValueOnce(3).mockResolvedValueOnce(0),
-      delete: jest.fn().mockResolvedValue(undefined),
-    };
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    jest.spyOn(ragService as any, 'getOrCreateTable').mockResolvedValue(mockTable);
-
-    const result = await ragService.deleteAllBySourceType('proj-1', 'code');
-
-    expect(result).toBe(3);
-    expect(mockTable.delete).toHaveBeenCalledWith("source = 'code'");
-  });
-
-  it('AC2: returns 0 when no records exist for the source type', async () => {
-    const ragService = new RagService(makeRagConfig());
-    const mockTable = {
-      countRows: jest.fn().mockResolvedValue(0),
-      query: jest.fn().mockReturnValue({
-        limit: jest.fn().mockReturnThis(),
-        toArray: jest.fn().mockResolvedValue([]),
-      }),
-      delete: jest.fn().mockResolvedValue(undefined),
-    };
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    jest.spyOn(ragService as any, 'getOrCreateTable').mockResolvedValue(mockTable);
-
-    const result = await ragService.deleteAllBySourceType('proj-1', 'code');
-
-    expect(result).toBe(0);
-  });
-
-  it('supports source-based deletes for the in-memory fallback table', async () => {
-    const mockEmbeddingService = {
-      embed: jest.fn().mockResolvedValue(new Float32Array(384).fill(0)),
-      providerName: 'ollama',
-      modelName: 'nomic-embed-text',
-      dimensions: 384,
-    };
-
-    const ragService = new RagService(makeRagConfig({ inMemoryOnly: true }), mockEmbeddingService as never);
-
-    await ragService.indexDocument('proj-1', {
-      source: 'code',
-      sourceId: 'node-1',
-      content: 'class ExampleNode',
-      metadata: {},
-    });
-    await ragService.indexDocument('proj-1', {
-      source: 'manual',
-      sourceId: 'doc-1',
-      content: 'manual ExampleDoc',
-      metadata: {},
-    });
-
-    const deleted = await ragService.deleteAllBySourceType('proj-1', 'code');
-    const remaining = await ragService.listDocuments('proj-1', 10);
-
-    expect(deleted).toBe(1);
-    expect(remaining).toHaveLength(1);
-    expect(remaining[0].source).toBe('manual');
-  });
-});
-
 describe('RagService.importGraphify (US-002)', () => {
   const mockRagConfig = makeRagConfig();
 
@@ -1046,8 +184,13 @@ describe('RagService.importGraphify (US-002)', () => {
     dimensions: 384,
   };
 
+  function makeRagService(): RagService {
+    const vectorStore = new VectorStore(mockRagConfig, mockEmbeddingService as never);
+    return new RagService(vectorStore);
+  }
+
   it('AC3: calls deleteAllBySourceType before indexing any nodes', async () => {
-    const ragService = new RagService(mockRagConfig, mockEmbeddingService as never);
+    const ragService = makeRagService();
     const callOrder: string[] = [];
     const deleteAllBySourceTypeSpy = jest.spyOn(ragService, 'deleteAllBySourceType').mockImplementation(async () => {
       callOrder.push('deleteAllBySourceType');
@@ -1070,7 +213,7 @@ describe('RagService.importGraphify (US-002)', () => {
   });
 
   it('AC4: with empty links, indexes each node with content "{type} {label} in {source_file}" (omitting source_file segment when absent)', async () => {
-    const ragService = new RagService(mockRagConfig, mockEmbeddingService as never);
+    const ragService = makeRagService();
     jest.spyOn(ragService, 'deleteAllBySourceType').mockResolvedValue(0);
     const indexDocumentSpy = jest.spyOn(ragService, 'indexDocument').mockResolvedValue(undefined);
 
@@ -1101,7 +244,7 @@ describe('RagService.importGraphify (US-002)', () => {
   });
 
   it('AC5: builds content that includes "{relation} {neighbor_label}" for each link where node is the source', async () => {
-    const ragService = new RagService(mockRagConfig, mockEmbeddingService as never);
+    const ragService = makeRagService();
     jest.spyOn(ragService, 'deleteAllBySourceType').mockResolvedValue(0);
     const indexDocumentSpy = jest.spyOn(ragService, 'indexDocument').mockResolvedValue(undefined);
 
@@ -1124,7 +267,7 @@ describe('RagService.importGraphify (US-002)', () => {
   });
 
   it('AC6: calls indexDocument with source: "code", sourceId: node.id, and metadata containing label, type, source_file, and community', async () => {
-    const ragService = new RagService(mockRagConfig, mockEmbeddingService as never);
+    const ragService = makeRagService();
     jest.spyOn(ragService, 'deleteAllBySourceType').mockResolvedValue(0);
     const indexDocumentSpy = jest.spyOn(ragService, 'indexDocument').mockResolvedValue(undefined);
 
@@ -1148,7 +291,7 @@ describe('RagService.importGraphify (US-002)', () => {
   });
 
   it('AC7: returns { imported: 0, cleared: N } for empty nodes array', async () => {
-    const ragService = new RagService(mockRagConfig, mockEmbeddingService as never);
+    const ragService = makeRagService();
     jest.spyOn(ragService, 'deleteAllBySourceType').mockResolvedValue(5);
 
     const result = await ragService.importGraphify('proj-1', [], []);
@@ -1157,7 +300,7 @@ describe('RagService.importGraphify (US-002)', () => {
   });
 
   it('defaults type to "node" when absent', async () => {
-    const ragService = new RagService(mockRagConfig, mockEmbeddingService as never);
+    const ragService = makeRagService();
     jest.spyOn(ragService, 'deleteAllBySourceType').mockResolvedValue(0);
     const indexDocumentSpy = jest.spyOn(ragService, 'indexDocument').mockResolvedValue(undefined);
 
@@ -1176,7 +319,7 @@ describe('RagService.importGraphify (US-002)', () => {
   });
 
   it('returns { imported: N, cleared: M } after successful import', async () => {
-    const ragService = new RagService(mockRagConfig, mockEmbeddingService as never);
+    const ragService = makeRagService();
     jest.spyOn(ragService, 'deleteAllBySourceType').mockResolvedValue(3);
     jest.spyOn(ragService, 'indexDocument').mockResolvedValue(undefined);
 
@@ -1188,195 +331,5 @@ describe('RagService.importGraphify (US-002)', () => {
     const result = await ragService.importGraphify('proj-1', nodes, []);
 
     expect(result).toEqual({ imported: 2, cleared: 3 });
-  });
-});
-
-describe('RagService.search — Provenance Envelope (AC-1 through AC-6)', () => {
-  function makeTableWithProvenance(allRows: AnyRecord[]) {
-    return {
-      countRows: jest.fn().mockResolvedValue(allRows.length || 1),
-      query: jest.fn().mockReturnValue({
-        limit: jest.fn().mockReturnThis(),
-        toArray: jest.fn().mockResolvedValue(allRows),
-      }),
-      search: jest.fn().mockResolvedValue(allRows),
-      vectorSearch: jest.fn().mockReturnValue({
-        distanceType: jest.fn().mockReturnThis(),
-        limit: jest.fn().mockReturnThis(),
-        toArray: jest.fn().mockResolvedValue(allRows),
-      }),
-    };
-  }
-
-  it('AC-1: SearchKbResponseDto.provenance is non-null on success', async () => {
-    const ragService = new RagService(
-      mockRagConfigForSearch,
-      mockEmbeddingServiceForSearch as never,
-    );
-
-    const doc = makeMockRecord('doc-1', 'authentication error content');
-    const mockTable = makeTableWithProvenance([doc]);
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    const result = await ragService.search('test-project', 'authentication error', 5);
-
-    expect(result.provenance).toBeDefined();
-    expect(result.provenance).not.toBeNull();
-  });
-
-  it('AC-2: SearchKbResponseDto.provenance.sources lists unique source_type + source_id pairs only once', async () => {
-    const ragService = new RagService(
-      mockRagConfigForSearch,
-      mockEmbeddingServiceForSearch as never,
-    );
-
-    // Multiple results with the same (source, source_id) pair
-    const createdAt = new Date().toISOString();
-    const doc1: AnyRecord = {
-      id: 'doc-1',
-      source: 'ticket',
-      source_id: 'ticket-123',
-      content: 'error in auth service',
-      vector: [],
-      metadata: '{}',
-      created_at: createdAt,
-      provider: 'ollama',
-      model: 'nomic-embed-text',
-    };
-    const doc2: AnyRecord = {
-      id: 'doc-2',
-      source: 'ticket',
-      source_id: 'ticket-123',
-      content: 'another error in auth service',
-      vector: [],
-      metadata: '{}',
-      created_at: createdAt,
-      provider: 'ollama',
-      model: 'nomic-embed-text',
-    };
-
-    const mockTable = makeTableWithProvenance([doc1, doc2]);
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    const result = await ragService.search('test-project', 'error auth', 5);
-
-    // Should have exactly one source entry despite multiple results from same source
-    const sourceEntry = result.provenance?.sources?.find(
-      (s) => s.sourceType === 'ticket' && s.sourceId === 'ticket-123',
-    );
-    expect(sourceEntry).toBeDefined();
-
-    // Count how many times this source appears in provenance.sources
-    const count = result.provenance?.sources?.filter(
-      (s) => s.sourceType === 'ticket' && s.sourceId === 'ticket-123',
-    ).length;
-    expect(count).toBe(1);
-  });
-
-  it('AC-3: KbResultDto.provenance.indexedAt is a valid ISO timestamp', async () => {
-    const ragService = new RagService(
-      mockRagConfigForSearch,
-      mockEmbeddingServiceForSearch as never,
-    );
-
-    const now = new Date().toISOString();
-    const doc = makeMockRecord('doc-1', 'test content');
-    doc.created_at = now;
-
-    const mockTable = makeTableWithProvenance([doc]);
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    const result = await ragService.search('test-project', 'test content', 5);
-
-    expect(result.results).toHaveLength(1);
-    const indexedAt = result.results[0].provenance?.indexedAt;
-    expect(indexedAt).toBeDefined();
-    // Check if it's a valid ISO string
-    expect(new Date(indexedAt as string).toISOString()).toBe(now);
-  });
-
-  it('AC-4: KbResultDto.provenance.sourceProjectId equals the request projectId', async () => {
-    const ragService = new RagService(
-      mockRagConfigForSearch,
-      mockEmbeddingServiceForSearch as never,
-    );
-
-    const doc = makeMockRecord('doc-1', 'test content');
-    const mockTable = makeTableWithProvenance([doc]);
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    const result = await ragService.search('test-project', 'test content', 5);
-
-    expect(result.results).toHaveLength(1);
-    expect(result.results[0].provenance?.sourceProjectId).toBe('test-project');
-  });
-
-  it('AC-5: No cross-project leakage — all results have sourceProjectId matching request', async () => {
-    const ragService = new RagService(
-      mockRagConfigForSearch,
-      mockEmbeddingServiceForSearch as never,
-    );
-
-    // Even if somehow a document from another project slipped in
-    const doc1 = makeMockRecord('doc-1', 'content from project-a');
-    const doc2 = makeMockRecord('doc-2', 'content from project-a');
-
-    const mockTable = makeTableWithProvenance([doc1, doc2]);
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_project-a', mockTable]]);
-
-    const result = await ragService.search('project-a', 'content', 5);
-
-    // All results must have sourceProjectId === 'project-a'
-    result.results.forEach((r) => {
-      expect(r.provenance?.sourceProjectId).toBe('project-a');
-    });
-  });
-
-  it('AC-6: SearchKbResponseDto.provenance.retrievedAt is within 1 second of server response', async () => {
-    const ragService = new RagService(
-      mockRagConfigForSearch,
-      mockEmbeddingServiceForSearch as never,
-    );
-
-    const doc = makeMockRecord('doc-1', 'test content');
-    const mockTable = makeTableWithProvenance([doc]);
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).lanceAvailable = true;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (ragService as any).tableCache = new Map([['project_test-project', mockTable]]);
-
-    const beforeSearch = Date.now();
-    const result = await ragService.search('test-project', 'test content', 5);
-    const afterSearch = Date.now();
-
-    const retrievedAt = result.provenance?.retrievedAt;
-    expect(retrievedAt).toBeDefined();
-
-    const retrievedAtMs = new Date(retrievedAt as string).getTime();
-    const timeDiff = Math.abs(retrievedAtMs - beforeSearch);
-
-    // Should be within 1 second (1000ms) of the search start
-    expect(timeDiff).toBeLessThan(1000);
   });
 });
